requirements.txt
eve
flask-swagger-ui
Eve-Swagger
gunicorn
bcrypt
